requirements.txt
django
appdirs==1.4.3
attrs==19.3.0
black==19.3b0
boltons==19.2.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
djangorestframework==3.10.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
oauthlib==3.1.0
pendulum==2.0.5
pika==1.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.0
python-dotenv==0.10.3
pytzdata==2019.3
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
selenium==3.141.0
six==1.12.0
toml==0.10.0
uritemplate==3.0.0
urllib3==1.25.6
dotenv
PyYAML
